requirements.txt
fastapi[all]
gunicorn
uvicorn
psycopg2-binary